Many changes: - improved keyboard input - added woflenstein textures - corrected walking on map - switched to full screen - increased size of map - code refactoring

+++ src/Player.h
@@ -25,8 +25,8 @@ private:
 	int PosOnMapX = 0;
 	int PosOnMapY = 0;
 	double angle = 0;
-	double toDiffX;
-	double toDiffY;
+	int toDiffX;
+	int toDiffY;
 	log4cpp::Category& logger = log4cpp::Category::getInstance("Player");
 public:
 	Player();
